Backend/rest_api/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
attrs==24.2.0
billiard==4.2.0
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.0
Django==5.0.9
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.2
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.66.0
grpcio-status==1.66.0
httplib2==0.22.0
idna==2.10
jeepney==0.8.0
kombu==5.4.0
mercadopago==2.2.3
msgpack==1.0.8
multidict==6.0.5
numpy==2.1.3
phonenumbers==8.13.37
pillow==10.3.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.2.0
redis==5.0.1
requests==2.25.1
rsa==4.9
scipy==1.14.1
SecretStorage==3.3.1
shapely==2.0.4
six==1.16.0
sqlparse==0.5.1
twilio==9.2.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.19
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
